GameObject position base

+++ Timer.cpp
@@ -5,16 +5,16 @@ Timer::Timer()
 	last = std::chrono::steady_clock::now();
 }
 
-float Timer::Mark()
+double Timer::Mark()
 {
 	const auto old = last;
 	last = std::chrono::steady_clock::now();
-	const std::chrono::duration<float> frameTime = last - old;
+	const std::chrono::duration<double> frameTime = last - old;
 
 	return frameTime.count();
 }
 
-float Timer::Peek()
+double Timer::Peek()
 {
-	return std::chrono::duration<float>(std::chrono::steady_clock::now() - last).count();
+	return std::chrono::duration<double>(std::chrono::steady_clock::now() - last).count();
 }
